# optimize memory usage in image load with Docker client integration