@nirs 2025-07-21
@yankay So with this we don't need to save the image before loading?

    podman save registry/repo/name:tag -o image.tar
    minikube load image.tar

This will not help the case when we ned to load into multiple clusters:

    podman save registry/repo/name:tag -o image.tar
    minikube load -p cluster1 image.tar
    minikube load -p cluster2 image.tar

Where the image is saved? how is it cleaned up?

Regardless, loading entire image to memory does not make sense in any case, so this looks like a good change.