@yankay 2025-07-22
> > @yankay So with this we don't need to save the image before loading?
> > ```
> > podman save registry/repo/name:tag -o image.tar
> > minikube load image.tar
> > ```
> > 
> > 
> >     
> >       
> >     
> > 
> >       
> >     
> > 
> >     
> >   
> > This will not help the case when we ned to load into multiple clusters:
> > ```
> > podman save registry/repo/name:tag -o image.tar
> > minikube load -p cluster1 image.tar
> > minikube load -p cluster2 image.tar
> > ```
> > 
> > 
> >     
> >       
> >     
> > 
> >       
> >     
> > 
> >     
> >   
> > Where the image is saved? how is it cleaned up?
> > Regardless, loading entire image to memory does not make sense in any case, so this looks like a good change.
> 
> HI @nirs
> 
> The image is saved at `~/.minikube/cache/images` , ref: https://github.com/kubernetes/minikube/blob/master/pkg/minikube/machine/cache_images.go#L247 .



> thank you @yankay do you mind putting a Before/After this PR in the description ? to proof the effectiveness ?

Thanks @medyagh 

Memory has indeed been reduced, but execution time has increased significantly. I'll look into why that is.


#### **Key Metrics**
| Metric                  | Original (`/usr/bin/minikube`) | Modified (`./out/minikube-linux-amd64`) | Change       |
|-------------------------|--------------------------------|----------------------------------------|--------------|
| **User Time**           | 81.23 sec                     | 79.57 sec                             | ↓ 2%        |
| **System Time**         | 5.29 sec                      | 23.22 sec                             | ↑ **339%**  |
| **Elapsed Time**        | 104.91 sec                    | 578.94 sec                            | ↑ **452%**  |
| **CPU Utilization**     | 82%                           | 17%                                   | ↓ 79%       |
| **Peak Memory (RSS)**   | 2,844 MB                      | 100 MB                                | ↓ **96.5%** |
| **Voluntary Context Switches** | 235,386                     | 2,120,505                             | ↑ **800%**  |



Logs:
```
# Before
root@gpu-3090:~# minikube version
minikube version: v1.36.0
commit: f8f52f5de11fc6ad8244afac475e1d0f96841df1-dirty

/usr/bin/time -v -o output1.log minikube image load elasticsearch:9.0.3  --daemon=true

root@gpu-3090:~# cat  output1.log
        Command being timed: "minikube image load elasticsearch:9.0.3 --daemon=true"
        User time (seconds): 81.23
        System time (seconds): 5.29
        Percent of CPU this job got: 82%
        Elapsed (wall clock) time (h:mm:ss or m:ss): 1:44.91
        Average shared text size (kbytes): 0
        Average unshared data size (kbytes): 0
        Average stack size (kbytes): 0
        Average total size (kbytes): 0
        Maximum resident set size (kbytes): 2844216
        Average resident set size (kbytes): 0
        Major (requiring I/O) page faults: 34
        Minor (reclaiming a frame) page faults: 809957
        Voluntary context switches: 235386
        Involuntary context switches: 356
        Swaps: 0
        File system inputs: 5848
        File system outputs: 1486840
        Socket messages sent: 0
        Socket messages received: 0
        Signals delivered: 0
        Page size (bytes): 4096
        Exit status: 0

# After
/usr/bin/time -v -o output2.log ./out/minikube-linux-amd64 image load elasticsearch:9.0.3  --daemon=true

root@gpu-3090:~/oss/minikube# cat output2.log
        Command being timed: "./out/minikube-linux-amd64 image load elasticsearch:9.0.3 --daemon=true"
        User time (seconds): 79.57
        System time (seconds): 23.22
        Percent of CPU this job got: 17%
        Elapsed (wall clock) time (h:mm:ss or m:ss): 9:38.94
        Average shared text size (kbytes): 0
        Average unshared data size (kbytes): 0
        Average stack size (kbytes): 0
        Average total size (kbytes): 0
        Maximum resident set size (kbytes): 102480
        Average resident set size (kbytes): 0
        Major (requiring I/O) page faults: 0
        Minor (reclaiming a frame) page faults: 17126
        Voluntary context switches: 2120505
        Involuntary context switches: 673
        Swaps: 0
        File system inputs: 0
        File system outputs: 1486888
        Socket messages sent: 0
        Socket messages received: 0
        Signals delivered: 0
        Page size (bytes): 4096
        Exit status: 0

  ```