@yankay 2025-07-23
> @yankay Results looks very good!
> 
> Can you compare with the split process?
> 
> ```
> podman save image -o image.tar
> minikube image load image.tar
> ```
> 
> We still have the issue of cleanup - keeping the image in the cache directory is not good, the user does not know that the image was added and nobody will clean it up. But I think this is not an issue in your change and it was already there.

Thanks, the performance information has been added to the PR description :-)