@nirs 2025-07-25
@medyagh this broke image load on macOS:

```console
% time minikube image load docker.io/library/alpine:latest -p c1

❌  Exiting due to GUEST_IMAGE_LOAD: Failed to load image: save to dir: caching images: caching image "/Users/nir/.minikube/cache/images/arm64/docker.io/library/alpine_latest": docker save: inspect image docker.io/library/alpine:latest via docker client: Cannot connect to the Docker daemon at unix:///var/run/docker.sock. Is the docker daemon running?

╭────────────────────────────────────────────────────────────────────────────────────────────────────────────────────────╮
│                                                                                                                        │
│    😿  If the above advice does not help, please let us know:                                                          │
│    👉  https://github.com/kubernetes/minikube/issues/new/choose                                                        │
│                                                                                                                        │
│    Please run `minikube logs --file=logs.txt` and attach logs.txt to the GitHub issue.                                 │
│    Please also attach the following file to the GitHub issue:                                                          │
│    - /var/folders/v1/lcvpk6v567x43cfsjqg_97t40000gn/T/minikube_image_dda93f623f5c27f0f7578776de8bc3f360c78630_0.log    │
│                                                                                                                        │
╰────────────────────────────────────────────────────────────────────────────────────────────────────────────────────────╯

minikube image load docker.io/library/alpine:latest -p c1  0.06s user 0.04s system 4% cpu 2.054 total
```

I guess this change introduces dependency on docker. Before this change we did not assume docker or podman client.

We don't have automated tests on macOS so this is not easy to test. To ensure we don't have regressions all changes must be tested on macOS.